backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.6.0
gunicorn==21.2.0
